Isolate peace communication code
