+++ package.json
@@ -129,6 +129,7 @@
     "detect-port": "^1.3.0",
     "electron": "^18.2.3",
     "electron-builder": "^23.0.3",
+    "electron-chromedriver": "^19.0.0",
     "electron-devtools-installer": "^3.2.0",
     "electron-notarize": "^1.2.1",
     "electron-rebuild": "^3.2.7",
@@ -160,6 +161,7 @@
     "rimraf": "^3.0.2",
     "sass": "^1.51.0",
     "sass-loader": "^12.6.0",
+    "selenium-webdriver": "^4.2.0",
     "style-loader": "^3.3.1",
     "terser-webpack-plugin": "^5.3.1",
     "ts-jest": "^28.0.2",
